Use sequencer in CitreaRollupBlueprint
