+++ bin/citrea/src/rollup/mod.rs
@@ -2,6 +2,7 @@ use async_trait::async_trait;
 pub use bitcoin::*;
 use citrea_fullnode::{CitreaFullnode, FullNode};
 use citrea_prover::{CitreaProver, Prover};
+use citrea_sequencer::{CitreaSequencer, Sequencer, SequencerConfig};
 pub use mock::*;
 use sov_modules_api::storage::HierarchicalStorageManager;
 use sov_modules_api::Spec;
@@ -17,6 +18,72 @@ mod mock;
 /// Overrides RollupBlueprint methods
 #[async_trait]
 pub trait CitreaRollupBlueprint: RollupBlueprint {
+    /// Creates a new sequencer
+    #[instrument(level = "trace", skip_all)]
+    async fn create_new_sequencer(
+        &self,
+        runtime_genesis_paths: &<Self::NativeRuntime as RuntimeTrait<
+            Self::NativeContext,
+            Self::DaSpec,
+        >>::GenesisPaths,
+        rollup_config: RollupConfig<Self::DaConfig>,
+        sequencer_config: SequencerConfig,
+    ) -> Result<Sequencer<Self>, anyhow::Error>
+    where
+        <Self::NativeContext as Spec>::Storage: NativeStorage,
+    {
+        let da_service = self.create_da_service(&rollup_config).await;
+
+        // TODO: Double check what kind of storage needed here.
+        // Maybe whole "prev_root" can be initialized inside runner
+        // Getting block here, so prover_service doesn't have to be `Send`
+
+        let ledger_db = self.create_ledger_db(&rollup_config);
+        let genesis_config = self.create_genesis_config(runtime_genesis_paths, &rollup_config)?;
+
+        let mut storage_manager = self.create_storage_manager(&rollup_config)?;
+        let prover_storage = storage_manager.create_finalized_storage()?;
+
+        let prev_root = ledger_db
+            .get_head_soft_batch()?
+            .map(|(number, _)| prover_storage.get_root_hash(number.0 + 1))
+            .transpose()?;
+
+        // TODO(https://github.com/Sovereign-Labs/sovereign-sdk/issues/1218)
+        let rpc_methods =
+            self.create_rpc_methods(&prover_storage, &ledger_db, &da_service, None)?;
+
+        let native_stf = StfBlueprint::new();
+
+        let genesis_root = prover_storage.get_root_hash(1);
+
+        let init_variant = match prev_root {
+            Some(root_hash) => InitVariant::Initialized(root_hash),
+            None => match genesis_root {
+                Ok(root_hash) => InitVariant::Initialized(root_hash),
+                _ => InitVariant::Genesis(genesis_config),
+            },
+        };
+
+        let seq = CitreaSequencer::new(
+            da_service,
+            prover_storage,
+            sequencer_config,
+            native_stf,
+            storage_manager,
+            init_variant,
+            rollup_config.public_keys,
+            ledger_db,
+            rollup_config.rpc,
+        )
+        .unwrap();
+
+        Ok(Sequencer {
+            runner: seq,
+            rpc_methods,
+        })
+    }
+
     /// Creates a new rollup.
     #[instrument(level = "trace", skip_all)]
     async fn create_new_rollup(
