+++ crates/sovereign-sdk/module-system/sov-modules-rollup-blueprint/Cargo.toml
@@ -1,46 +1,36 @@
 [package]
 name = "sov-modules-rollup-blueprint"
-description = "This crate contains abstractions needed to create a new rollup"
+version = { workspace = true }
 authors = { workspace = true }
 edition = { workspace = true }
 homepage = { workspace = true }
 license = { workspace = true }
-repository = { workspace = true }
-version = { workspace = true }
-readme = "README.md"
-resolver = "2"
 publish = true
+readme = "README.md"
+repository = { workspace = true }
+resolver = "2"
+description = "This crate contains abstractions needed to create a new rollup"
 
 [dependencies]
-sov-rollup-interface = { path = "../../rollup-interface", features = [
-    "native",
-] }
-sov-stf-runner = { path = "../../full-node/sov-stf-runner", features = [
-    "native",
-] }
-citrea-sequencer = { path = "../../../sequencer", features = [] }
-sov-state = { path = "../sov-state" }
 sequencer-client = { path = "../../../sequencer-client" }
-sov-modules-api = { path = "../../module-system/sov-modules-api", features = [
-    "native",
-] }
 sov-cli = { path = "../../module-system/sov-cli" }
+sov-modules-api = { path = "../../module-system/sov-modules-api", features = ["native"] }
+sov-rollup-interface = { path = "../../rollup-interface", features = ["native"] }
+sov-sequencer = { path = "../../full-node/sov-sequencer" }
+sov-state = { path = "../sov-state" }
+sov-stf-runner = { path = "../../full-node/sov-stf-runner", features = ["native"] }
 
-sov-modules-stf-blueprint = { path = "../../module-system/sov-modules-stf-blueprint", features = [
-    "native",
-] }
 sov-db = { path = "../../full-node/db/sov-db" }
+sov-modules-stf-blueprint = { path = "../../module-system/sov-modules-stf-blueprint", features = ["native"] }
 
-sov-ledger-rpc = { path = "../../full-node/sov-ledger-rpc", features = [
-    "server",
-] }
+sov-ledger-rpc = { path = "../../full-node/sov-ledger-rpc", features = ["server"] }
 
-hex = { workspace = true }
 anyhow = { workspace = true }
+async-trait = { workspace = true }
+borsh = { workspace = true }
+hex = { workspace = true }
+jsonrpsee = { workspace = true, features = ["http-client", "server"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
-async-trait = { workspace = true }
-jsonrpsee = { workspace = true, features = ["http-client", "server"] }
 tokio = { workspace = true }
 tracing = { workspace = true }
-borsh = { workspace = true }
