+++ crates/sovereign-sdk/module-system/sov-modules-rollup-blueprint/src/lib.rs
@@ -6,7 +6,6 @@ mod wallet;
 use std::net::SocketAddr;
 
 use async_trait::async_trait;
-use citrea_sequencer::{CitreaSequencer, SequencerConfig};
 pub use runtime_rpc::*;
 use sov_db::ledger_db::LedgerDB;
 use sov_modules_api::{Context, DaSpec, Spec};
@@ -126,150 +125,6 @@ pub trait RollupBlueprint: Sized + Send + Sync {
         LedgerDB::with_path(&rollup_config.storage.path).expect("Ledger DB failed to open")
     }
 
-    /// Creates a new sequencer
-    #[instrument(level = "trace", skip_all)]
-    async fn create_new_sequencer(
-        &self,
-        runtime_genesis_paths: &<Self::NativeRuntime as RuntimeTrait<
-            Self::NativeContext,
-            Self::DaSpec,
-        >>::GenesisPaths,
-        rollup_config: RollupConfig<Self::DaConfig>,
-        sequencer_config: SequencerConfig,
-    ) -> Result<Sequencer<Self>, anyhow::Error>
-    where
-        <Self::NativeContext as Spec>::Storage: NativeStorage,
-    {
-        let da_service = self.create_da_service(&rollup_config).await;
-
-        // TODO: Double check what kind of storage needed here.
-        // Maybe whole "prev_root" can be initialized inside runner
-        // Getting block here, so prover_service doesn't have to be `Send`
-
-        let ledger_db = self.create_ledger_db(&rollup_config);
-        let genesis_config = self.create_genesis_config(runtime_genesis_paths, &rollup_config)?;
-
-        let mut storage_manager = self.create_storage_manager(&rollup_config)?;
-        let prover_storage = storage_manager.create_finalized_storage()?;
-
-        let prev_root = ledger_db
-            .get_head_soft_batch()?
-            .map(|(number, _)| prover_storage.get_root_hash(number.0 + 1))
-            .transpose()?;
-
-        // TODO(https://github.com/Sovereign-Labs/sovereign-sdk/issues/1218)
-        let rpc_methods =
-            self.create_rpc_methods(&prover_storage, &ledger_db, &da_service, None)?;
-
-        let native_stf = StfBlueprint::new();
-
-        let genesis_root = prover_storage.get_root_hash(1);
-
-        let init_variant = match prev_root {
-            Some(root_hash) => InitVariant::Initialized(root_hash),
-            None => match genesis_root {
-                Ok(root_hash) => InitVariant::Initialized(root_hash),
-                _ => InitVariant::Genesis(genesis_config),
-            },
-        };
-
-        let seq = CitreaSequencer::new(
-            da_service,
-            prover_storage,
-            sequencer_config,
-            native_stf,
-            storage_manager,
-            init_variant,
-            rollup_config.public_keys,
-            ledger_db,
-            rollup_config.rpc,
-        )
-        .unwrap();
-
-        Ok(Sequencer {
-            runner: seq,
-            rpc_methods,
-        })
-    }
-
-    /// Creates a new prover
-    #[instrument(level = "trace", skip_all)]
-    async fn create_new_prover(
-        &self,
-        runtime_genesis_paths: &<Self::NativeRuntime as RuntimeTrait<
-            Self::NativeContext,
-            Self::DaSpec,
-        >>::GenesisPaths,
-        rollup_config: RollupConfig<Self::DaConfig>,
-        prover_config: ProverConfig,
-    ) -> Result<Prover<Self>, anyhow::Error>
-    where
-        <Self::NativeContext as Spec>::Storage: NativeStorage,
-    {
-        let da_service = self.create_da_service(&rollup_config).await;
-
-        let prover_service = self
-            .create_prover_service(prover_config.clone(), &rollup_config, &da_service)
-            .await;
-
-        // TODO: Double check what kind of storage needed here.
-        // Maybe whole "prev_root" can be initialized inside runner
-        // Getting block here, so prover_service doesn't have to be `Send`
-
-        let ledger_db = self.create_ledger_db(&rollup_config);
-        let genesis_config = self.create_genesis_config(runtime_genesis_paths, &rollup_config)?;
-
-        let mut storage_manager = self.create_storage_manager(&rollup_config)?;
-        let prover_storage = storage_manager.create_finalized_storage()?;
-
-        let prev_root = ledger_db
-            .get_head_soft_batch()?
-            .map(|(number, _)| prover_storage.get_root_hash(number.0 + 1))
-            .transpose()?;
-
-        let runner_config = rollup_config.runner.expect("Runner config is missing");
-        // TODO(https://github.com/Sovereign-Labs/sovereign-sdk/issues/1218)
-        let rpc_methods = self.create_rpc_methods(
-            &prover_storage,
-            &ledger_db,
-            &da_service,
-            Some(runner_config.sequencer_client_url.clone()),
-        )?;
-
-        let native_stf = StfBlueprint::new();
-
-        let genesis_root = prover_storage.get_root_hash(1);
-
-        let init_variant = match prev_root {
-            Some(root_hash) => InitVariant::Initialized(root_hash),
-            None => match genesis_root {
-                Ok(root_hash) => InitVariant::Initialized(root_hash),
-                _ => InitVariant::Genesis(genesis_config),
-            },
-        };
-
-        let code_commitment = self.get_code_commitment();
-
-        let runner = StateTransitionRunner::new(
-            runner_config,
-            rollup_config.public_keys,
-            rollup_config.rpc,
-            da_service,
-            ledger_db,
-            native_stf,
-            storage_manager,
-            init_variant,
-            Some(prover_service),
-            Some(prover_config),
-            code_commitment,
-        )?;
-
-        Ok(Prover {
-            runner,
-            rpc_methods,
-        })
-    }
-
     /// Creates a new rollup.
     #[instrument(level = "trace", skip_all)]
     async fn create_new_rollup(
@@ -344,49 +199,6 @@ pub trait RollupBlueprint: Sized + Send + Sync {
     }
 }
 
-/// Sequencer stf runner
-pub struct Sequencer<S: RollupBlueprint> {
-    /// The State Transition Runner of Sequencer.
-    #[allow(clippy::type_complexity)]
-    pub runner: CitreaSequencer<
-        S::NativeContext,
-        S::DaService,
-        S::StorageManager,
-        S::Vm,
-        StfBlueprint<S::NativeContext, S::DaSpec, S::Vm, S::NativeRuntime>,
-    >,
-    /// Rpc methods for the rollup.
-    pub rpc_methods: jsonrpsee::RpcModule<()>,
-}
-
-impl<S: RollupBlueprint> Sequencer<S> {
-    /// Runs the sequencer.
-    #[instrument(
-        name = "Sequencer",
-        level = "info",
-        skip_all,
-        err,
-        ret(level = "error")
-    )]
-    pub async fn run(self) -> Result<(), anyhow::Error> {
-        self.run_and_report_rpc_port(None).await
-    }
-
-    /// Runs the sequencer.
-    pub async fn run_and_report_rpc_port(
-        self,
-        channel: Option<oneshot::Sender<SocketAddr>>,
-    ) -> Result<(), anyhow::Error> {
-        let mut seq = self.runner;
-        seq.start_rpc_server(channel, self.rpc_methods)
-            .instrument(tracing::Span::current())
-            .await
-            .unwrap();
-        seq.run().await?;
-        Ok(())
-    }
-}
-
 /// Dependencies needed to run the rollup.
 pub struct FullNode<S: RollupBlueprint> {
     /// The State Transition Runner.
@@ -437,54 +249,6 @@ impl<S: RollupBlueprint> FullNode<S> {
             .await;
 
         runner.run_in_process().await?;
-        Ok(())
-    }
-}
-
-/// Dependencies needed to run the rollup.
-pub struct Prover<S: RollupBlueprint> {
-    /// The State Transition Runner.
-    #[allow(clippy::type_complexity)]
-    pub runner: StateTransitionRunner<
-        StfBlueprint<S::NativeContext, S::DaSpec, S::Vm, S::NativeRuntime>,
-        S::StorageManager,
-        S::DaService,
-        S::Vm,
-        S::ProverService,
-        S::NativeContext,
-    >,
-    /// Rpc methods for the rollup.
-    pub rpc_methods: jsonrpsee::RpcModule<()>,
-}
-
-impl<S: RollupBlueprint> Prover<S> {
-    /// Runs the rollup.
-    #[instrument(name = "Prover", level = "info", skip_all, err, ret(level = "error"))]
-    pub async fn run(self) -> Result<(), anyhow::Error> {
-        self.run_and_report_rpc_port(None).await
-    }
-
-    /// Only run the rpc.
-    pub async fn run_rpc(self) -> Result<(), anyhow::Error> {
-        self.runner
-            .start_rpc_server(self.rpc_methods, None)
-            .instrument(tracing::Span::current())
-            .await;
-        Ok(())
-    }
-
-    /// Runs the rollup. Reports rpc port to the caller using the provided channel.
-    pub async fn run_and_report_rpc_port(
-        self,
-        channel: Option<oneshot::Sender<SocketAddr>>,
-    ) -> Result<(), anyhow::Error> {
-        let mut runner = self.runner;
-        runner
-            .start_rpc_server(self.rpc_methods, channel)
-            .instrument(tracing::Span::current())
-            .await;
-
-        runner.run_prover_process().await?;
         Ok(())
     }
 }
